+++ server/src/core/server/graph/mutators/index.ts
@@ -2,6 +2,7 @@ import GraphContext from "coral-server/graph/context";
 
 import { Actions } from "./Actions";
 import { Comments } from "./Comments";
+import { DSAReports } from "./DSAReports";
 import { Redis } from "./Redis";
 import { Settings } from "./Settings";
 import { Sites } from "./Sites";
@@ -11,6 +12,7 @@ import { Users } from "./Users";
 const root = (ctx: GraphContext) => ({
   Actions: Actions(ctx),
   Comments: Comments(ctx),
+  DSAReports: DSAReports(ctx),
   Settings: Settings(ctx),
   Stories: Stories(ctx),
   Users: Users(ctx),
